+++ src/include/duckdb/execution/operator/persistent/batch_sink_helper.hpp
@@ -1,0 +1,149 @@
+//===----------------------------------------------------------------------===//
+//                         DuckDB
+//
+// duckdb/execution/operator/persistent/batch_sink_helper.hpp
+//
+//
+//===----------------------------------------------------------------------===//
+
+#pragma once
+
+#include "duckdb/common/common.hpp"
+#include "duckdb/storage/temporary_memory_manager.hpp"
+#include "duckdb/parallel/interrupt.hpp"
+#include "duckdb/storage/buffer_manager.hpp"
+
+namespace duckdb {
+
+class BatchSinkHelper {
+public:
+	BatchSinkHelper(ClientContext &context, idx_t initial_memory_request) :
+			context(context), unflushed_memory_usage(0), min_batch_index(0), can_increase_memory(true) {
+		memory_state = TemporaryMemoryManager::Get(context).Register(context);
+		SetMemorySize(initial_memory_request);
+	}
+
+private:
+	ClientContext &context;
+	//! Temporary memory state
+	unique_ptr<TemporaryMemoryState> memory_state;
+	//! Total memory usage of unflushed rows
+	atomic<idx_t> unflushed_memory_usage;
+	//! Minimum batch size
+	atomic<idx_t> min_batch_index;
+	//! The available memory for unflushed rows
+	atomic<idx_t> available_memory;
+	//! Blocked task lock
+	mutex blocked_task_lock;
+	//! The set of blocked tasks
+	vector<InterruptState> blocked_tasks;
+	//! Whether or not we can request additional memory
+	bool can_increase_memory;
+
+public:
+	void SetMemorySize(idx_t size) {
+		// request at most 1/4th of all available memory
+		idx_t total_max_memory = BufferManager::GetBufferManager(context).GetQueryMaxMemory();
+		idx_t request_cap = total_max_memory / 4;
+
+		size = MinValue<idx_t>(size, request_cap);
+		if (size <= available_memory) {
+			return;
+		}
+
+		memory_state->SetRemainingSize(context, size);
+		auto next_reservation = memory_state->GetReservation();
+		if (available_memory >= next_reservation) {
+			// we tried to ask for more memory but were declined
+			// stop asking for more memory
+			can_increase_memory = false;
+		}
+		available_memory = next_reservation;
+	}
+
+	void IncreaseMemory() {
+		if (!can_increase_memory) {
+			return;
+		}
+		SetMemorySize(available_memory * 2);
+	}
+
+	bool OutOfMemory(idx_t batch_index) {
+#ifdef DUCKDB_ALTERNATIVE_VERIFY
+		// alternative verify - always report that we are out of memory to test this code path
+		return true;
+#else
+		if (unflushed_memory_usage >= available_memory) {
+			lock_guard<mutex> l(blocked_task_lock);
+			if (batch_index > min_batch_index) {
+				// exceeded available memory and we are not the minimum batch index- try to increase it
+				IncreaseMemory();
+				if (unflushed_memory_usage >= available_memory) {
+					// STILL out of memory
+					return true;
+				}
+			}
+		}
+		return false;
+#endif
+	}
+
+	void BlockTask(InterruptState &state) {
+		blocked_tasks.push_back(state);
+	}
+
+	bool UnblockTasks() {
+		lock_guard<mutex> l(blocked_task_lock);
+		return UnblockTasksInternal();
+	}
+
+	bool UnblockTasksInternal() {
+		if (blocked_tasks.empty()) {
+			return false;
+		}
+		for (auto &entry : blocked_tasks) {
+			entry.Callback();
+		}
+		blocked_tasks.clear();
+		return true;
+	}
+
+	void UpdateMinBatchIndex(idx_t current_min_batch_index) {
+		if (min_batch_index >= current_min_batch_index) {
+			return;
+		}
+		lock_guard<mutex> l(blocked_task_lock);
+		auto new_batch_index = MaxValue<idx_t>(min_batch_index, current_min_batch_index);
+		if (new_batch_index != min_batch_index) {
+			// new batch index! unblock all tasks
+			min_batch_index = new_batch_index;
+			UnblockTasksInternal();
+		}
+	}
+
+	idx_t AvailableMemory() const {
+		return available_memory;
+	}
+
+	idx_t GetMinimumBatchIndex() const {
+		return min_batch_index;
+	}
+
+	mutex &GetBlockedTaskLock() {
+		return blocked_task_lock;
+	}
+
+	void IncreaseUnflushedMemory(idx_t memory_increase) {
+		unflushed_memory_usage += memory_increase;
+	}
+
+	void ReduceUnflushedMemory(idx_t memory_reduction) {
+		unflushed_memory_usage -= memory_reduction;
+	}
+
+	bool IsMinimumBatchIndex(idx_t batch_index) {
+		return batch_index <= min_batch_index;
+	}
+};
+
+} // namespace duckdb
